% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awapi.R
\name{aw_24_hrs_condition}
\alias{aw_24_hrs_condition}
\title{Get 24 hour current conditions for a given location. The first part of the function gets the location key and passes it through the url containing the data}
\usage{
aw_24_hrs_condition(
  lat,
  lon,
  tz = "Europe/London",
  apikey = Sys.getenv("awapi_key")
)
}
\arguments{
\item{lat}{Latitude of location.}

\item{lon}{Longitude of location.}

\item{tz}{Default timezone is Europe/London. See http://www.timezoneconverter.com/cgi-bin/zoneinfo.tzc to find your timezone.}

\item{apikey}{apikey needed to access data from the site.}
}
\value{
List of 24 hour current conditions for the location.
}
\description{
Get 24 hour current conditions for a given location. The first part of the function gets the location key and passes it through the url containing the data
}
