% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awapi.R
\name{aw_check_status_code}
\alias{aw_check_status_code}
\title{Check the status code for your apikey}
\usage{
aw_check_status_code(lat, lon, apikey = Sys.getenv("awapi_key"))
}
\arguments{
\item{lat}{Latitude of location.}

\item{lon}{Longitude of location.}

\item{apikey}{apikey needed to access data from the site.}
}
\value{
message signifying status of api connection
}
\description{
Check the status code for your apikey
}
