% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awapi.R
\name{aw_is_api_key_set}
\alias{aw_is_api_key_set}
\title{Check if the api key has been added to environment variables}
\usage{
aw_is_api_key_set()
}
\value{
Returns TRUE if the api key has been set in the "ACCUWEATHERAPIKEY" environment variable, otherwise FALSE.
}
\description{
Check if the api key has been added to environment variables
}
