% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awapi.R
\name{aw_set_api_key}
\alias{aw_set_api_key}
\title{Sets the Accu Weather API Key.}
\usage{
aw_set_api_key(x)
}
\arguments{
\item{x}{Your AccuWeather API key.}
}
\value{

}
\description{
Sets the Accu Weather API Key.
}
