% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_converter.R
\name{C_to_F}
\alias{C_to_F}
\title{Converts temperature from one unit to another}
\usage{
C_to_F(temp)
}
\arguments{
\item{temp}{Temperature in Celsius}
}
\value{
R object or list to be converted to Fahrenheit
}
\description{
Converts temperature from one unit to another
}
