% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_converter.R
\name{F_to_K}
\alias{F_to_K}
\title{Converts temperature from one unit to another}
\usage{
F_to_K(temp)
}
\arguments{
\item{temp}{Temperature in Fahrenheit}
}
\value{
R object or list to be converted to Kelvin
}
\description{
Converts temperature from one unit to another
}
