% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_converter.R
\name{K_to_C}
\alias{K_to_C}
\title{Converts temperature from one unit to another}
\usage{
K_to_C(temp)
}
\arguments{
\item{temp}{Temperature in Kelvin.}
}
\value{
R object or list to be converted to Celsius
}
\description{
Converts temperature from one unit to another
}
